Changing the ball to a Rect and a circle so it can be modified.Starting on the ball movement function.

+++ Paddle.py
@@ -5,15 +5,15 @@ from Config import Config
 
 class Paddle:
 	def __init__(self, game_display):
-			self.x = int(Config['game']['display_width'] * 0.45)
-			self.y = int(Config['game']['display_height'] * 0.92)
+			self.x = Config['game']['display_width'] * 0.45
+			self.y = Config['game']['display_height'] * 0.92
 			self.game_display=game_display
 			self.paddle_width = Config['paddle']['paddle_width']
 			self.paddle_height = Config['paddle']['paddle_height']
 	
 			
 	def draw(self):
-		pygame.draw.rect(self.game_display, Config['colors']['red'], [self.x, self.y, self.paddle_width, self.paddle_height ])
+		pygame.draw.rect(self.game_display, Config['colors']['red'], [self.x, self.y, self.paddle_width, self.paddle_height])
 		
 	def movement(self, x_change):
 		self.x += x_change
